+++ src/main/java/com/back/domain/news/common/service/AnalysisNewsService.java
@@ -26,12 +26,7 @@ import java.util.concurrent.ExecutionException;
 @RequiredArgsConstructor
 public class AnalysisNewsService {
 
-    private final AiService aiService;
-    private final ObjectMapper objectMapper;
-    private final RateLimiter rateLimiter;
-
-    @Qualifier("bucket")
-    private final Bucket bucket;
+    private final NewsAnalysisBatchService newsAnalysisBatchService;
 
     @Value("${news.filter.batch.size:2}") // 배치 크기 설정, 3으로 하면 본문 길면 깨짐
     private int batchSize;
@@ -57,7 +52,7 @@ public class AnalysisNewsService {
 
         // 비동기 작업들 시작
         List<CompletableFuture<Void>> futures = batches.stream()
-                .map(batch -> processBatchAsync(batch)
+                .map(batch -> newsAnalysisBatchService.processBatchAsync(batch)
                         .thenAccept(result -> {
                             // 완료되는 대로 즉시 결과에 추가
                             allResults.addAll(result);
@@ -84,32 +79,6 @@ public class AnalysisNewsService {
 
         log.info("뉴스 필터링 완료 - 최종 결과: {}개", allResults.size());
         return new ArrayList<>(allResults);
-    }
-
-    @Async("newsExecutor")
-    public CompletableFuture<List<AnalyzedNewsDto>> processBatchAsync(List<RealNewsDto> batch) {
-
-        log.info("스레드: {}, 배치 시작 시간: {}",
-                Thread.currentThread().getName(), System.currentTimeMillis());
-        try {
-            // Rate limiting - 토큰 얻을 때까지 계속 시도
-            rateLimiter.waitForRateLimit();
-            log.debug("배치 처리 시작 - 기사 수: {}", batch.size());
-
-            NewsAnalysisProcessor processor = new NewsAnalysisProcessor(batch, objectMapper);
-            List<AnalyzedNewsDto> result = aiService.process(processor);
-
-            log.debug("배치 처리 완료 - 결과 수: {}", result.size());
-
-            log.info("스레드: {}, 배치 완료 시간: {}",
-                    Thread.currentThread().getName(), System.currentTimeMillis());
-
-            return CompletableFuture.completedFuture(result);
-
-        } catch (Exception e) {
-            log.error("배치 {}개 처리 중 오류 발생", batch.size(), e);
-            return CompletableFuture.completedFuture(List.of()); // AI 처리 실패시만 빈 리스트
-        }
     }
 
 }
