+++ src/main/resources/application.yml
@@ -78,7 +78,7 @@ naver:
   client-secret: ${NAVER_CLIENT_SECRET}
   base-url: "https://openapi.naver.com/v1/search/news?query="
   news:
-    display: 4
+    display: 2
     sort: sim
   crawling:
     delay: 1000 # 줄이지 말아주세요
